Sensible default for publicPath
I am not at all a webpack expert, so this change may not be appropriate. However when writing my own config, I did not need a public path, default being '/'.
So I think we should rely on this default when computing the component path (but once again I am not a webpack expert).